+++ lib/views/settings_screen/components/import_field.dart
@@ -210,6 +210,8 @@ class _MyDropRegionState extends State<MyDropRegion> with Logger, TickerProvider
 
   Future<void> _processFile(String filename, String content) async {
     final settingsRepo = TomlSerializableRepository(path.join(appDataPath, "Settings.toml"), Settings.fromJson);
+    // Ensure settings file exists
+    await getIt<SettingsManager>().save();
 
     if (!filename.toLowerCase().endsWith(".toml")) {
       logWarning("Filename $filename does not end with .toml, trying to use anyway.");
@@ -234,6 +236,7 @@ class _MyDropRegionState extends State<MyDropRegion> with Logger, TickerProvider
 
     setState(() => _isDragOver = false);
     // BuildContextAbstractor is not available, so neither is showUserDialog
+    if (!mounted) return;
     await Navigator.of(context, rootNavigator: true).push(PhotoBoothDialogPage(
       child: Padding(
         padding: const EdgeInsets.all(32),
@@ -251,7 +254,7 @@ class _MyDropRegionState extends State<MyDropRegion> with Logger, TickerProvider
         }, updates: updates)),
       ),
       barrierDismissible: true,
-    ).createRoute(context));
+    ).createRoute());
   }
 
   void _onDropLeave(DropEvent event) {